requirements.txt
alembic==1.8.1
aniso8601==9.0.1
apispec==6.0.2
bidict==0.22.1
blinker==1.5
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
flask-apispec==0.11.4
Flask-Cors==3.0.10
Flask-Gunicorn==0.1.1
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
Flask-Migrate==4.0.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
gunicorn==20.1.0
faker==18.6.0
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
packaging==21.3
psycopg2-binary==2.9.5
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.6
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.44
urllib3==1.26.13
uvicorn==0.20.0
waitress==2.1.2
webargs==8.2.0
Werkzeug==2.2.2
zipp==3.10.0
